requirements.txt
openai
langchain
langchain_community
newsapi-python